Add ImageChanged event handling for plugins
Updated the `MainForm` class to include an event handler for the `ImageChanged` event of the `plugin` object. A new method, `PluginControl_ImageChanged`, has been added to update the plugin icon in the UI when requested. This method iterates through the controls in `flowLayoutPanel1` to find and update the corresponding plugin icon with the new image.